@hemartin-at-redhat 2022-05-12
I completely changed to regex, times are not relevant, since they change between 1.5 and 2 for both implementations. I just switched the compute time to another code path. However I think the implementation with Regexp is a little more comprehensible.